fix the password display incorrectly
